ENH: Add abstract integer types for use with image size
Add SizeValueType, IdentifierType, IndexValueType and OffsetValueType
as typedefs available in the top level itk namespace.

Use these instead of "long" or "unsigned long" to achieve 64-bit
image size capability on Win64.

Replace "long" and "unsigned long" with appropriate abstract types
everywhere in the code where they are presently used for one of the
above purposes. Other uses of long remain where their use does not
require a 64-bit value space.

+++ Code/Algorithms/itkDeformableSimplexMesh3DFilter.h
@@ -127,11 +127,11 @@ public:
   typedef typename InputMeshType::NeighborListType           InputNeighbors;
   typedef typename InputMeshType::NeighborListType::iterator InputNeighborsIterator;
 
-  typedef itk::MapContainer< unsigned long, std::set< unsigned long > > VertexNeighborListType;
-  typedef std::set< unsigned long >                                     NeighborSetType;
-  typedef std::set< unsigned long >                                     IndexSetType;
-  typedef typename NeighborSetType::iterator                            NeighborSetIterator;
-  typedef typename IndexSetType::iterator                               IndexSetIterator;
+  typedef itk::MapContainer< size_t, std::set< size_t > > VertexNeighborListType;
+  typedef std::set< size_t >                              NeighborSetType;
+  typedef std::set< size_t >                              IndexSetType;
+  typedef typename NeighborSetType::iterator              NeighborSetIterator;
+  typedef typename IndexSetType::iterator                 IndexSetIterator;
 
   typedef typename InputMeshType::GeometryMapType GeometryMapType;
   typedef typename GeometryMapType::Pointer       GeometryMapPointer;
